src/store-front/requirements.txt
Flask==1.1.2
python-dotenv
requests==2.25.1
